+++ src/exceptiongroup/_catch.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import inspect
 import sys
 from collections.abc import Callable, Iterable, Mapping
 from contextlib import AbstractContextManager
@@ -49,9 +50,15 @@ class _Catcher:
             matched, excgroup = excgroup.split(exc_types)
             if matched:
                 try:
-                    handler(matched)
+                    result = handler(matched)
                 except BaseException as new_exc:
                     new_exceptions.append(new_exc)
+                else:
+                    if inspect.iscoroutine(result):
+                        raise TypeError(
+                            f"Error trying to handle {matched!r} with {handler!r}. "
+                            "Exception handler must be a sync function."
+                        ) from exc
 
             if not excgroup:
                 break